+++ frontend/src/layout/ErrorBoundary/index.ts
@@ -1,0 +1,1 @@
+export { ErrorBoundary } from './ErrorBoundary'
